+++ packages/express/test/unit/build.test.ts
@@ -63,7 +63,16 @@ const readDirectoryRecursively = (dirPath: string, basePath = ''): string[] => {
   return files;
 };
 
-const fixtures = {
+const fixtures: Record<
+  string,
+  {
+    handler: string[];
+    moduleType: 'cjs' | 'esm';
+    projectSettings?: {
+      outputDirectory?: string;
+    };
+  }
+> = {
   '01-index-js-no-module': {
     handler: ['index.js'],
     moduleType: 'cjs',
@@ -112,8 +121,37 @@ const fixtures = {
     handler: ['index.mjs'],
     moduleType: 'esm',
   },
-  '13-app-js-no-module': {
+  '13-index-ts-output-directory': {
+    handler: ['dist', 'index.js'],
+    moduleType: 'esm',
+    projectSettings: {
+      outputDirectory: 'dist',
+    },
+  },
+  '14-app-js-no-module': {
     handler: ['app.js'],
+    moduleType: 'cjs',
+  },
+  '15-src-server-js-no-module': {
+    handler: ['src', 'server.js'],
+    moduleType: 'cjs',
+  },
+  '16-src-app-js-no-module': {
+    handler: ['src', 'app.js'],
+    moduleType: 'cjs',
+  },
+  '17-multiple-matches': {
+    // matches alphabetically first
+    handler: ['src', 'app.js'],
+    moduleType: 'cjs',
+  },
+  '18-multiple-matches-with-no-exp': {
+    // src/app.js is alphabetically first, but its contents don't match the regex
+    handler: ['src', 'index.js'],
+    moduleType: 'cjs',
+  },
+  '19-index-cts-module-tsconfig': {
+    handler: ['index.cjs'],
     moduleType: 'cjs',
   },
 };
@@ -131,7 +169,13 @@ describe('build', () => {
       const result = await build({
         files,
         workPath,
-        config,
+        config: {
+          ...config,
+          projectSettings: {
+            ...config.projectSettings,
+            ...fixtureConfig.projectSettings,
+          },
+        },
         meta,
         // Entrypoint is just used as the BOA function name
         entrypoint: 'this value is not used',
@@ -151,7 +195,7 @@ describe('build', () => {
       } else {
         throw new Error('entrypoint is not defined');
       }
-    });
+    }, 10000);
   }
 });
 describe('failing fixtures', () => {